Inline location option mapping
